[src/vasprintf.c] Fixed several bugs in buffer_sandwich, which could yield memory corruption with non-default memory allocators and other undefined behavior; check the return value when this function is called.
Details:
  * check integer overflow on the size computations;
  * computation of q and r without tests;
  * fixed the case where r > len (e.g. len = 1 and tz = 1);
  * in the loop, increase the str pointer only when needed, otherwise
    str could be beyond of the end of the string of the digits, which
    is undefined behavior.

The first consequence of the r > len issue was an incorrect generated
string. Moreover, since the generated string was shorter than expected,
mpfr_free_str would provide an incorrect buffer size to the "free"
function of the current GMP memory allocator. By default, this size is
ignored, but it may matter if the memory allocators have been changed
with the mp_set_memory_functions GMP function, in which case a possible
consequence could be memory corruption.

git-svn-id: svn://scm.gforge.inria.fr/svn/mpfr/trunk@12869 280ebfd0-de03-0410-8827-d642c229c3f4