[tests] In C++, do not cast to "void *" in the tests of macros.
These casts make the build of these tests fail with a C++ compiler.
But this means that the use of macros will not introduce any issue
concerning the type of the arguments, i.e. with a C++ compiler, the
tests are still complete after this change.

However, GCC's -Werror=c++-compat still cannot be used.
To avoid this issue, diagnostic pragmas will be needed.

git-svn-id: svn://scm.gforge.inria.fr/svn/mpfr/trunk@14460 280ebfd0-de03-0410-8827-d642c229c3f4