[tests/{tset_si.c,tset_sj.c}] Fix when -DMPFR_USE_NO_MACRO is used.
The mpfr_get_exp() function checks that the argument is a regular
number, but the corresponding macro doesn't. The test code was
calling mpfr_get_exp() on 0, thus failed with MPFR_USE_NO_MACRO
defined. Here, the code was correct with the macro, because the
value was not used for the particular case 0.

This is fixed by testing 0 earlier (and doing a goto, but the code
has fewer tests and should be simpler to understand). Moreover,
mpfr_get_exp has been replaced by MPFR_GET_EXP, so that the argument
is checked with --enable-assert.

git-svn-id: svn://scm.gforge.inria.fr/svn/mpfr/trunk@14464 280ebfd0-de03-0410-8827-d642c229c3f4